+++ src/main/java/ru_hogwarts_school/service/StudentService.java
@@ -1,10 +1,14 @@
 package ru_hogwarts_school.service;
 
 import org.springframework.stereotype.Service;
+import ru_hogwarts_school.exception.FacultyNotFoundException;
+import ru_hogwarts_school.exception.StudentNotFoundException;
+import ru_hogwarts_school.model.Faculty;
 import ru_hogwarts_school.model.Student;
 import ru_hogwarts_school.repository.StudentRepository;
 
 import java.util.Collection;
+import java.util.Optional;
 
 @Service
 public class StudentService {
@@ -17,24 +21,37 @@ public class StudentService {
 
 
     public Student addStudent(Student student) {
+        student.setId(null);
         return studentRepository.save(student);
     }
 
 
     public Student getStudent(long id) {
-        return studentRepository.findById(id).get();
+        Optional<Student> student = studentRepository.findById(id);
+        if (student.isPresent()) {
+            return studentRepository.findById(id).get();
+        }
+        throw new StudentNotFoundException();
     }
 
 
     public Student editStudent(Student student) {
-        return studentRepository.save(student);
+        Optional<Student> stu = studentRepository.findById(student.getId());
+        if (stu.isPresent()) {
+            return studentRepository.save(student);
+        }
+        throw new StudentNotFoundException();
     }
 
 
-    public void deleteStudent(long id) {
-        studentRepository.deleteById(id);
+    public Student deleteStudent(long id) {
+        Optional<Student> student = studentRepository.findById(id);
+        if (student.isPresent()) {
+            studentRepository.deleteById(id);
+            return student.get();
+        }
+        throw new StudentNotFoundException();
     }
-
 
     public Collection<Student> getAllStudent() {
         return studentRepository.findAll();
@@ -42,8 +59,22 @@ public class StudentService {
 
 
     public Collection<Student> getAllStudentByAge(int age) {
+        Collection<Student> students = studentRepository.getAllStudentByAge(age);
+        if (students.isEmpty()) {
+            throw new StudentNotFoundException();
+        }
         return studentRepository.getAllStudentByAge(age);
     }
 
+    public Collection<Student> findByAgeBetween(int min, int max) {
+        Collection<Student> students = studentRepository.findByAgeBetween(min, max);
+        if (students.isEmpty()) {
+            throw new StudentNotFoundException();
+        }
+        return studentRepository.findByAgeBetween(min, max);
+    }
 
+    public Faculty findFacultyFromStudent(long id) {
+        return getStudent(id).getFaculty();
+    }
 }
